+++ HLTrigger/Configuration/python/HLT_75e33/modules/hltHgcalSoALayerClustersProducerHSi_cfi.py
@@ -8,8 +8,8 @@ hltHgcalSoALayerClustersProducerHSi = cms.EDProducer("HGCalSoALayerClustersProdu
     hgcalRecHitsLayerClustersSoA = cms.InputTag("hltHgcalSoARecHitsLayerClustersProducerHSi"),
     hgcalRecHitsSoA = cms.InputTag("hltHgcalSoARecHitsProducerHSi"),
     detector = cms.string('FH'),
-    positionDeltaRho2 = cms.double(1.69),
-    thresholdW0 = cms.double(2.9)
+    positionDeltaRho2 = cms.float(1.69),
+    thresholdW0 = cms.float(2.9)
 )
 
 hltHgcalSoALayerClustersProducerHSiSerialSync = makeSerialClone(hltHgcalSoALayerClustersProducerHSi,
