+++ HLTrigger/Configuration/python/HLT_75e33/modules/hltHgcalSoARecHitsLayerClustersProducerL1Seeded_cfi.py
@@ -6,7 +6,7 @@ hltHgcalSoARecHitsLayerClustersProducerL1Seeded = cms.EDProducer("HGCalCLUEsteri
     ),
     hgcalRecHitsSoA = cms.InputTag("hltHgcalSoARecHitsProducerL1Seeded"),
     detector = cms.string('EE'),
-    deltac = cms.double(1.3),
-    kappa = cms.double(9),
-    outlierDeltaFactor = cms.double(2.0)
+    deltac = cms.float(1.3),
+    kappa = cms.float(9),
+    outlierDeltaFactor = cms.float(2.0)
 )
